+++ packages/core/include/biosim/core/grid.h
@@ -1,7 +1,7 @@
 /*
  * HOST-ONLY: this header uses a heap pointer and standard library types.
  * Do NOT include it from OpenCL kernel sources (.cl files).
- * GPU kernels receive the raw uint16_t * buffer as a kernel argument.
+ * GPU kernels receive the raw uint32_t * buffer as a kernel argument.
  */
 #ifndef BIOSIM_CORE_GRID_H
 #define BIOSIM_CORE_GRID_H
@@ -13,24 +13,24 @@
 
 /*
  * Flat 2D grid buffer, row-major: cells[y * size_x + x].
- * Cell encoding: BIOSIM_GRID_EMPTY (0), BIOSIM_GRID_BARRIER (0xFFFF),
- * or a 1-based agent index in [1, 0xFFFE].
+ * Cell encoding: BIOSIM_GRID_EMPTY (0), BIOSIM_GRID_BARRIER (0xFFFFFFFF),
+ * or a 1-based agent index in [1, 0xFFFFFFFE].
  */
 typedef struct {
-    uint16_t *cells;
-    int16_t size_x;
-    int16_t size_y;
+    uint32_t *cells;
+    int32_t size_x;
+    int32_t size_y;
 } biosim_grid_t;
 
 /* Lifecycle */
-biosim_status_t biosim_grid_create(int16_t size_x, int16_t size_y, biosim_grid_t *out);
+biosim_status_t biosim_grid_create(int32_t size_x, int32_t size_y, biosim_grid_t *out);
 void biosim_grid_free(biosim_grid_t *grid);
 void biosim_grid_zero_fill(biosim_grid_t *grid);
 
 /* Bounds check and raw cell access — at/set assert in_bounds */
 bool biosim_grid_in_bounds(const biosim_grid_t *grid, biosim_coord_t coord);
-uint16_t biosim_grid_at(const biosim_grid_t *grid, biosim_coord_t coord);
-void biosim_grid_set(biosim_grid_t *grid, biosim_coord_t coord, uint16_t value);
+uint32_t biosim_grid_at(const biosim_grid_t *grid, biosim_coord_t coord);
+void biosim_grid_set(biosim_grid_t *grid, biosim_coord_t coord, uint32_t value);
 
 /* Cell predicates */
 bool biosim_grid_is_empty(const biosim_grid_t *grid, biosim_coord_t coord);
@@ -42,9 +42,9 @@ bool biosim_grid_is_occupied(const biosim_grid_t *grid, biosim_coord_t coord);
  * Out-of-bounds cells are silently skipped. The callback receives the
  * coordinate and current cell value; pass arbitrary context via `sim`.
  */
-typedef void (*biosim_grid_visitor_t)(biosim_coord_t coord, uint16_t cell, void *sim);
+typedef void (*biosim_grid_visitor_t)(biosim_coord_t coord, uint32_t cell, void *sim);
 void biosim_grid_visit_neighborhood(const biosim_grid_t *grid, biosim_coord_t center,
-                                    int16_t radius, biosim_grid_visitor_t visitor, void *sim);
+                                    int32_t radius, biosim_grid_visitor_t visitor, void *sim);
 
 /*
  * Find a random empty cell by random probing then linear scan fallback.
